+++ frontend/src/components/budget-overview/MainCategoryView.tsx
@@ -4,7 +4,7 @@ import SubCategoryView from "./SubCategoryView";
 import "./styles/MainCategoryView.css"
 import {Position} from "../../model/Position";
 import {useEffect, useState} from "react";
-import {getMainSum} from "../../service/utils/sumHelpers";
+import {getMainSum} from "../../service/utils/accountingHelpers";
 import {formatMoney} from "../../service/utils/formattingHelpers";
 
 type MainCategoryViewProps = {
