+++ frontend/src/components/event-overview/MainCategoryOverview.tsx
@@ -2,7 +2,7 @@ import {MainCategory} from "../../model/MainCategory";
 import {SubCategory} from "../../model/SubCategory";
 import {Position} from "../../model/Position";
 import {formatMoney} from "../../service/utils/formattingHelpers";
-import {getMainSum} from "../../service/utils/sumHelpers";
+import {getMainSum} from "../../service/utils/accountingHelpers";
 import SubCategoryOverview from "./SubCategoryOverview";
