+++ frontend/src/pages/BudgetOverview.tsx
@@ -1,6 +1,6 @@
 import EventDetailView from "../components/budget-overview/EventDetailView";
 import MainCategoryView from "../components/budget-overview/MainCategoryView";
-import {getBalance} from "../service/utils/sumHelpers";
+import {getBalance} from "../service/utils/accountingHelpers";
 import BalanceView from "../components/budget-overview/BalanceView";
 import "./styles/BudgetOverview.css"
 import {MainCategory} from "../model/MainCategory";
